chore: Refactored events into seperate directory
