+++ src/events/on_application_command_error.py
