+++ src/events/on_command_error.py
