+++ src/events/on_raw_reaction_remove.py
